# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The reply mode/tone
 */
export type GenerateReplyBodyMode =
  (typeof GenerateReplyBodyMode)[keyof typeof GenerateReplyBodyMode];

export const GenerateReplyBodyMode = {
  romantic: "romantic",
  funny: "funny",
  savage: "savage",
  emotional: "emotional",
} as const;

export interface GenerateReplyBody {
  /** The conversation text to analyze */
  conversation: string;
  /** The reply mode/tone */
  mode: GenerateReplyBodyMode;
  /** Reply language (e.g. english, spanish, french) */
  lang?: string;
}

export type GenerateReplyResponseMood = {
  flirty: number;
  playful: number;
  tension: number;
  warmth: number;
};

export interface GenerateReplyResponse {
  /** Three reply variants */
  variants: string[];
  mood: GenerateReplyResponseMood;
  /** Interest level 0-100 */
  interestLevel: number;
  signals: string[];
}

export interface RizzScoreBody {
  opener: string;
}

export interface RizzScoreResponse {
  score: number;
  grade: string;
  verdict: string;
  pros: string[];
  cons: string[];
  improved: string;
}

export interface TrackerAdviceBody {
  name: string;
  trend: number[];
  note?: string;
}

export interface TrackerAdviceResponse {
  advice: string;
  nextMove: string;
  greenFlag: boolean;
}

export interface SavedReply {
  id: number;
  conversationSnippet: string;
  mode: string;
  reply: string;
  interestLevel: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}