# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: replybro
    description: ReplyBro reply generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /replybro/generate:
    post:
      operationId: generateReply
      tags: [replybro]
      summary: Generate AI reply variants with mood analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReplyBody"
      responses:
        "200":
          description: Generated reply variants with mood analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateReplyResponse"

  /replybro/rizz:
    post:
      operationId: scoreRizz
      tags: [replybro]
      summary: Score a rizz opening line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RizzScoreBody"
      responses:
        "200":
          description: Rizz score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RizzScoreResponse"

  /replybro/advice:
    post:
      operationId: getTrackerAdvice
      tags: [replybro]
      summary: Get AI coaching for relationship tracker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackerAdviceBody"
      responses:
        "200":
          description: AI relationship advice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerAdviceResponse"

  /replybro/history:
    get:
      operationId: listReplyHistory
      tags: [replybro]
      summary: List saved reply history
      responses:
        "200":
          description: List of saved replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedReply"

  /replybro/history/{id}:
    delete:
      operationId: deleteReplyHistory
      tags: [replybro]
      summary: Delete a saved reply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"

  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"

  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    GenerateReplyBody:
      type: object
      properties:
        conversation:
          type: string
          description: The conversation text to analyze
        mode:
          type: string
          enum: [romantic, funny, savage, emotional]
          description: The reply mode/tone
        lang:
          type: string
          description: Reply language (e.g. english, spanish, french)
          default: english
      required:
        - conversation
        - mode

    GenerateReplyResponse:
      type: object
      properties:
        variants:
          type: array
          items:
            type: string
          description: Three reply variants
        mood:
          type: object
          properties:
            flirty:
              type: integer
            playful:
              type: integer
            tension:
              type: integer
            warmth:
              type: integer
          required: [flirty, playful, tension, warmth]
        interestLevel:
          type: integer
          description: Interest level 0-100
        signals:
          type: array
          items:
            type: string
      required:
        - variants
        - mood
        - interestLevel
        - signals

    RizzScoreBody:
      type: object
      properties:
        opener:
          type: string
      required:
        - opener

    RizzScoreResponse:
      type: object
      properties:
        score:
          type: integer
        grade:
          type: string
        verdict:
          type: string
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        improved:
          type: string
      required: [score, grade, verdict, pros, cons, improved]

    TrackerAdviceBody:
      type: object
      properties:
        name:
          type: string
        trend:
          type: array
          items:
            type: integer
        note:
          type: string
      required: [name, trend]

    TrackerAdviceResponse:
      type: object
      properties:
        advice:
          type: string
        nextMove:
          type: string
        greenFlag:
          type: boolean
      required: [advice, nextMove, greenFlag]

    SavedReply:
      type: object
      properties:
        id:
          type: integer
        conversationSnippet:
          type: string
        mode:
          type: string
        reply:
          type: string
        interestLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationSnippet
        - mode
        - reply
        - interestLevel
        - createdAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error