# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate AI reply variants with mood analysis
 */
export const generateReplyBodyLangDefault = `english`;

export const GenerateReplyBody = zod.object({
  conversation: zod.string().describe("The conversation text to analyze"),
  mode: zod
    .enum(["romantic", "funny", "savage", "emotional"])
    .describe("The reply mode\/tone"),
  lang: zod
    .string()
    .default(generateReplyBodyLangDefault)
    .describe("Reply language (e.g. english, spanish, french)"),
});

export const GenerateReplyResponse = zod.object({
  variants: zod.array(zod.string()).describe("Three reply variants"),
  mood: zod.object({
    flirty: zod.number(),
    playful: zod.number(),
    tension: zod.number(),
    warmth: zod.number(),
  }),
  interestLevel: zod.number().describe("Interest level 0-100"),
  signals: zod.array(zod.string()),
});

/**
 * @summary Score a rizz opening line
 */
export const ScoreRizzBody = zod.object({
  opener: zod.string(),
});

export const ScoreRizzResponse = zod.object({
  score: zod.number(),
  grade: zod.string(),
  verdict: zod.string(),
  pros: zod.array(zod.string()),
  cons: zod.array(zod.string()),
  improved: zod.string(),
});

/**
 * @summary Get AI coaching for relationship tracker
 */
export const GetTrackerAdviceBody = zod.object({
  name: zod.string(),
  trend: zod.array(zod.number()),
  note: zod.string().optional(),
});

export const GetTrackerAdviceResponse = zod.object({
  advice: zod.string(),
  nextMove: zod.string(),
  greenFlag: zod.boolean(),
});

/**
 * @summary List saved reply history
 */
export const ListReplyHistoryResponseItem = zod.object({
  id: zod.number(),
  conversationSnippet: zod.string(),
  mode: zod.string(),
  reply: zod.string(),
  interestLevel: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListReplyHistoryResponse = zod.array(ListReplyHistoryResponseItem);

/**
 * @summary Delete a saved reply
 */
export const DeleteReplyHistoryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all conversations
 */
export const ListAnthropicConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicConversationsResponse = zod.array(
  ListAnthropicConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateAnthropicConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnthropicConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListAnthropicMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAnthropicMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAnthropicMessagesResponse = zod.array(
  ListAnthropicMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendAnthropicMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendAnthropicMessageBody = zod.object({
  content: zod.string(),
});